+++ backend/src/controllers/trending.controllers.js
@@ -1,7 +1,7 @@
 // Pull in required modules
 const trendingService = require("../service/trending.service");
 const { handleApiError } = require("../utils/apiUtils");
-const parseIntWithFallback = require("../utils/number");
+const { parseIntWithFallback } = require("../utils/validation");
 
 async function getTrending(req, res, next) {
   try {
