+++ backend/src/routes/users.routes.js
@@ -6,7 +6,8 @@ const router = Router();
 router.post('/login', authController.loginUser);
 router.post('/register', authController.registerUser);
 
-// Test route to verify authentication middleware
+// Test route to verify authentication middleware - this only returns if we successfully authenticated the logged in user using their access_token.
+// This method will be removed after additional protected endpoints are added.
 router.get('/protected', authenticate, (req, res) => {
   res.json({ message: 'You are authenticated!', user: req.user });
 });
