+++ backend/src/utils/apiUtils.js
@@ -2,23 +2,27 @@ const supabase = require('../utils/supabaseClient');
 
 async function authenticate(req, res, next) {
 
+  // Try to get our auth token - if we didn't get on then return 401.
   const authHeader = req.headers.authorization;
-
   if (!authHeader || !authHeader.startsWith('Bearer ')) {
     return res.status(401).json({ error: 'No token provided.' });
   }
 
+  // Extract our auth bearer token
   const token = authHeader.split(' ')[1];
   
+
   try {
     const { data, error } = await supabase.auth.getUser(token);
     if (error || !data.user) {
       return res.status(401).json({ error: 'Invalid or expired token.' });
     }
-    req.user = data.user; // Attach user info to request
+    
+    // Attach user info to request
+    req.user = data.user; 
     next();
   } catch (err) {
-    return res.status(500).json({ error: 'Authentication failed.' });
+    handleApiError(err, res, next);
   }
 }
