--- backend/src/utils/number.js
@@ -1,6 +1,0 @@
-function parseIntWithFallback(v, fallback) {
-  const n = parseInt(v, 10);
-  return Number.isNaN(n) ? fallback : n;
-}  
-
-module.exports = parseIntWithFallback
